StdIn now support files from network
I mean kind of... It's not make stream from network but
download it first to local than open the file with std::ifstream

This also means adding +1 dependency to project: cpp-httplib
It's header-only but suppport for https scheme requires some
dependencies make it work.

This increased compile-time, but make the thing done. I may be
revisit this decision later.